tests/ui/requirements.txt
pytest>=7.0
Pillow>=9.0
pytest-html>=3.0
pytest-timeout>=2.0
